+++ CHANGELOG.md
@@ -69,6 +69,7 @@ Compatibility:
 * Support writing to `RData.dfree` for native extensions (#2830, #2732, #2165, @eregon).
 * Fix `IO#write` and support multiple arguments with different encodings (#2829, @andrykonchin).
 * Fix `Array` methods `reject`, `reject!`, `inject`, `map`, `select`, `each_index` and handle a case when array is modified by a passed block like CRuby does (#2822, andrykonchin, @eregon).
+* Fix `EncodingError` exception message when Symbol has invalid encoding (#2850, @andrykonchin).
 
 Performance: